Move SelectedTab.tsx into a folder
