+++ src/lib/Components/Artist/ArtistConsignButton.tsx
@@ -6,7 +6,7 @@ import { useTracking } from "react-tracking"
 import styled from "styled-components/native"
 
 import { ArtistConsignButton_artist } from "__generated__/ArtistConsignButton_artist.graphql"
-import { TabName, useSelectedTab } from "lib/NativeModules/SelectedTab"
+import { TabName, useSelectedTab } from "lib/NativeModules/SelectedTab/SelectedTab"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
 import { Router } from "lib/utils/router"
 import { Schema } from "lib/utils/track"
